switch from filepath to virtualfile and mark it transient
neither filepath nor virtualfile can be serialized, hence the transient.
we only need the workspace while the build is running, once it is finished
we grab the screenshot from the artifact storage and don't need the
workspace anymore.

virtualfile allows us to transparently handle local and remote files so
we can access them from remote slave builds as well.